+++ qml/AddressDelegate.qml
@@ -7,46 +7,38 @@ import QtQuick.Controls.Material 2.3
 import QtQuick.Layouts 1.12
 
 ItemDelegate {
-    id: self
-
     required property Address address
-
+    id: self
     hoverEnabled: true
-    leftPadding: 0
-    rightPadding: 0
-    topPadding: 0
-    bottomPadding: 0
-    background: Item {}
-
-    spacing: 8
-
+    padding: 0
+    background: null
+    spacing: constants.p2
     contentItem: RowLayout {
-        spacing: 16
-
-        Label {
-            Layout.maximumWidth: 200
+        spacing: constants.p3
+        CopyableLabel {
+            Layout.fillWidth: true
+            Layout.leftMargin: constants.p1
             elide: Text.ElideRight
             text: address.data["address"]
-            font.pixelSize: 12
-            font.capitalization: Font.AllUppercase
+            font.pixelSize: 14
             font.styleName: 'Regular'
-            opacity: 0.6
         }
-
         Label {
             text: address.data["address_type"]
-            font.pixelSize: 12
+            horizontalAlignment: Text.AlignHCenter
+            font.pixelSize: 14
             font.capitalization: Font.AllUppercase
             font.styleName: 'Regular'
-            opacity: 0.6
+            Layout.minimumWidth: 50
         }
-
         Label {
             text: address.data["tx_count"]
-            font.pixelSize: 12
+            horizontalAlignment: Label.AlignHCenter
+            font.pixelSize: 14
             font.capitalization: Font.AllUppercase
             font.styleName: 'Regular'
-            opacity: 0.6
+            Layout.minimumWidth: 50
+            Layout.rightMargin: constants.p1
         }
     }
 }